requirements.txt
pandas
numpy
scikit-learn
joblib
flask
click
pyyaml
streamlit
matplotlib
seaborn
requests
pytest
